# Optimize decryptObjectInfo