--- a/go.mod
+++ b/go.mod
@@ -90,3 +90,7 @@ require (
 	gopkg.in/olivere/elastic.v5 v5.0.86
 	gopkg.in/yaml.v2 v2.2.8
 )
+
+replace (
+	github.com/minio/sio => ../sio
+)