… cmd/testdata/decryptObjectInfo.json.zst
… go.sum