## Motivation and Context

`decryptObjectInfo` is a significant bottleneck when listing objects.

Reduce the allocations for a significant speedup. This PR covers the basic optimizations.

The MB/s is the number of decrypted objects per second.

Requires:

https://github.com/minio/sio/pull/40

`go.mod` replace term until upstream is available.

```
λ benchcmp before.txt after.txt
benchmark                          old ns/op     new ns/op     delta
Benchmark_decryptObjectInfo-32     24260928      808656        -96.67%

benchmark                          old MB/s     new MB/s     speedup
Benchmark_decryptObjectInfo-32     0.04         1.24         31.00x

benchmark                          old allocs     new allocs     delta
Benchmark_decryptObjectInfo-32     75112          48996          -34.77%

benchmark                          old bytes     new bytes     delta
Benchmark_decryptObjectInfo-32     287694772     4228076       -98.53%
```

## How to test this PR?

Tests/benchmarks are included.

## Types of changes
- [x] Optimization.