@klauspost 2020-10-26
@aead If you can tag sio 0.2.1 I can update the `go.mod`