@minio-trusted 2020-10-27
```
Error running mint automation
+ mkdir -p 10726-7394d0b/gopath/src/github.com/minio
+ git -C 10726-7394d0b/gopath/src/github.com/minio clone --quiet https://github.com/minio/minio.git
+ git -C 10726-7394d0b/gopath/src/github.com/minio/minio remote add minio https://github.com/minio/minio.git
+ git -C 10726-7394d0b/gopath/src/github.com/minio/minio fetch --quiet minio pull/10726/head:pr10726
+ git -C 10726-7394d0b/gopath/src/github.com/minio/minio checkout --quiet pr10726
+ GO111MODULE=on
+ GOPATH=/home/mint-auto/mint-auto/10726-7394d0b/gopath
+ make -C 10726-7394d0b/gopath/src/github.com/minio/minio --quiet
go: github.com/minio/sio@v0.2.1: parsing ../sio/go.mod: open /home/mint-auto/mint-auto/10726-7394d0b/gopath/src/github.com/minio/sio/go.mod: no such file or directory
go: github.com/minio/sio@v0.2.1: parsing ../sio/go.mod: open /home/mint-auto/mint-auto/10726-7394d0b/gopath/src/github.com/minio/sio/go.mod: no such file or directory
make: *** [build] Error 1

```